Moved montage utilities out of tasks and removed unused functions
