+++ ramutils/tasks/__init__.py
@@ -8,7 +8,7 @@ from ._wrapper import memory, task, make_task
 from .classifier import *
 from .events import *
 from .misc import *
-from .montage import *
 from .odin import *
 from .powers import *
 from .summary import *
+from .montage import *
